sys-kernel: Unify kernel config between coreos-kernel and coreos-modules
The kernel config can affect btf generation, and we need btf information
to match between the two builds otherwise modules don't load. Accomplish
that by including an empty user specified initramfs in the vmlinux built
by coreos-modules. This vmlinux file is only used as a base for btf
generation, and is discarded after the build.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>